refine uts
